requirements.txt
pycdr==0.1.5
eclipse-zenoh
Django